charms/sabnzbd-k8s/pyproject.toml
[project]
name = "sabnzbd-k8s"
dynamic = ["version"]
requires-python = ">=3.12"

dependencies = [
    "ops>=2.17",
    "charmarr-lib-core",
    "charmarr-lib-vpn",
    "pydantic>=2.0",
    "lightkube>=0.15",
    "lightkube-extensions",
    "httpx>=0.27",
    "charmed-service-mesh-helpers",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "codespell>=2.0",
    "pyright>=1.1",
    "pytest>=8.0",
    "coverage[toml]>=7.0",
    "ops[testing]>=2.17",
]
integration = [
    "charmarr-lib-testing>=0.4.0",
    "pytest-jubilant",
    "pytest-bdd>=7.0",
    "jubilant>=1.6",
    "lightkube>=0.15",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "D",    # pydocstyle - no docstrings required in tests
]

[tool.ruff.lint.isort]
known-first-party = ["charm", "charmarr_lib"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pyright]
extraPaths = ["src", "lib"]
pythonVersion = "3.12"
pythonPlatform = "Linux"
typeCheckingMode = "standard"
include = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --tb=short -x"
log_cli_level = "WARNING"
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/**"]

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[dependency-groups]
dev = [
    "tox>=4.32.0",
]
